Extract FontManager
